+++ toro-exoplayer/src/main/java/im/ene/toro/exoplayer/ExoPlayable.java
@@ -126,13 +126,11 @@ public class ExoPlayable extends PlayableImpl {
       if (selector instanceof DefaultTrackSelector) {
         MappedTrackInfo trackInfo = ((DefaultTrackSelector) selector).getCurrentMappedTrackInfo();
         if (trackInfo != null) {
-          if (trackInfo.getTrackTypeRendererSupport(C.TRACK_TYPE_VIDEO)
-              == RENDERER_SUPPORT_UNSUPPORTED_TRACKS) {
+          if (trackInfo.getTypeSupport(C.TRACK_TYPE_VIDEO) == RENDERER_SUPPORT_UNSUPPORTED_TRACKS) {
             onErrorMessage(toro.getString(R.string.error_unsupported_video));
           }
 
-          if (trackInfo.getTrackTypeRendererSupport(C.TRACK_TYPE_AUDIO)
-              == RENDERER_SUPPORT_UNSUPPORTED_TRACKS) {
+          if (trackInfo.getTypeSupport(C.TRACK_TYPE_AUDIO) == RENDERER_SUPPORT_UNSUPPORTED_TRACKS) {
             onErrorMessage(toro.getString(R.string.error_unsupported_audio));
           }
         }